duplicated component removed
