+++ src/components/location-search/index.tsx
@@ -47,6 +47,8 @@ function LocationSearchComponent({
     setDropdownVisible(true);
   };
 
+  const width = 'w-40 sm:w-72 lg:w-96 xl:w-[620px]';
+
   return (
     <div
       className={cn({
@@ -56,8 +58,9 @@ function LocationSearchComponent({
     >
       <div
         className={cn({
-          'z-40 flex h-14 w-40 items-center rounded-t-[4px] border border-secondary-500 bg-brand-50 px-4 font-inter leading-4 sm:w-72 lg:w-96 xl:w-[620px]':
+          'relative z-40 flex h-14 items-center rounded-t-[4px] border border-secondary-500 bg-brand-50 px-4 font-inter leading-4':
             true,
+          [width]: true,
           'rounded-[4px]': !locationSearch,
           'rounded-t-[4px]': locationSearch,
           'w-[calc(100vw-34px-40px)]': isMobile,
@@ -84,7 +87,14 @@ function LocationSearchComponent({
       {(isLoading && isFetching) ||
         (dropdownVisible && locationSearch && !!OPTIONS.length && (
           <div className="relative">
-            <div className="absolute right-0 top-0 z-50 w-40 flex-1 rounded-b-[4px] border-b border-l border-r border-secondary-500 bg-brand-50 px-10 font-inter leading-4 text-secondary-700 shadow-lg md:w-96 lg:w-[620px]">
+            <div
+              className={cn({
+                'absolute right-0 top-0 z-50 flex-1 rounded-b-[4px] border-b border-l border-r border-secondary-500 bg-brand-50 px-10 font-inter leading-4 text-secondary-700 shadow-lg':
+                  true,
+                [width]: true,
+                'left-0 right-3 w-[calc(100vw-34px-40px)]': isMobile,
+              })}
+            >
               {isLoading && isFetching && <Loading />}
               {locationSearch && OPTIONS.length > 0 && (
                 <ul id="location-options" role="listbox" className="space-y-2 py-2">
