+++ src/components/monitors/card/index.tsx
@@ -64,7 +64,7 @@ const MonitorCard: FC<Partial<Monitor> & { color?: string }> = (monitor) => {
 
       <div className="flex justify-end">
         <div className="relative h-[40px] w-[40px] sm:h-[60px] sm:w-[60px]">
-          <Image src={`/svgs/theme-icons/${theme?.toLowerCase()}.svg`} fill alt={theme} />
+          <Image src={`/svgs/theme-icons/${theme?.toLowerCase()}.svg`} fill alt={theme || id} />
         </div>
       </div>
     </div>
